Some update
In repository layer change Generic type of id from Long to Integer, because in entity CheckMetadata type of id is Integer

+++ src/main/java/space/eliseev/iplatformmoex/repository/CheckMetadataRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import space.eliseev.iplatformmoex.model.entity.CheckMetadata;
 
 @Repository
-public interface CheckMetadataRepository extends JpaRepository<CheckMetadata, Long> {
+public interface CheckMetadataRepository extends JpaRepository<CheckMetadata, Integer> {
 
 }
